根据OpenTCS文件，对Dto类稍作修改； 添加opentcs与DTO数据对应表文件。

+++ code/mj-tcs/tcs-web/src/main/java/com/mj/tcs/api/v1/dto/PathDto.java
@@ -57,11 +57,11 @@ public class PathDto extends BaseEntityDto {
 
     @DtoField
     @Column
-    private int maxVelocity;
+    private double maxVelocity;
 
     @DtoField
     @Column
-    private int maxReverseVelocity;
+    private double maxReverseVelocity;
 
     @DtoField
     @Column
@@ -122,7 +122,7 @@ public class PathDto extends BaseEntityDto {
         this.routingCost = routingCost;
     }
 
-    public int getMaxVelocity() {
+    public double getMaxVelocity() {
         return maxVelocity;
     }
 
@@ -130,7 +130,7 @@ public class PathDto extends BaseEntityDto {
         this.maxVelocity = maxVelocity;
     }
 
-    public int getMaxReverseVelocity() {
+    public double getMaxReverseVelocity() {
         return maxReverseVelocity;
     }
